refactor(tests): split tests/collection_cli_truth.rs by command into tests/cli_collection_truth_*.rs
Splits the 3058-LOC `tests/collection_cli_truth.rs` (which exceeded the
1500-LOC budget for `tests/*.rs` files) into ten per-command files plus
a shared fixtures helper, applying the same per-feature split rule used
for the source-file migrations. Tests move verbatim — only `use` paths
and `mod common;` declarations were adjusted.

Per-command files (each ≤ 1500-LOC budget):
- tests/cli_collection_truth_add.rs           — `quaid collection add`
- tests/cli_collection_truth_audit.rs         — `quaid collection audit`
- tests/cli_collection_truth_info.rs          — `quaid collection info`
- tests/cli_collection_truth_list.rs          — `quaid collection list`
- tests/cli_collection_truth_migrate_uuids.rs — `quaid collection migrate-uuids`
- tests/cli_collection_truth_quarantine.rs    — `quaid collection quarantine`
- tests/cli_collection_truth_remap.rs         — `quaid collection remap`
- tests/cli_collection_truth_restore.rs       — `quaid collection restore`
- tests/cli_collection_truth_slug_routing.rs  — slug-routing semantics
- tests/cli_collection_truth_sync.rs          — `quaid collection sync`

Shared fixtures: tests/common/truth_fixtures.rs.

Removed: tests/collection_cli_truth.rs (3058 LOC).

Verification:
  baseline (parent ab64567):  passed=1357 failed=0
  this commit:                passed=1357 failed=0

Both clippy channels green:
  cargo clippy --all-targets --locked -- -D warnings
  cargo clippy --all-targets --no-default-features \
    --features bundled,online-model --locked -- -D warnings

Part of OpenSpec change \`extract-inline-tests-to-integration\`.
